requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
mysql-connector-python==8.0.26
bcrypt==3.2.0
python-dotenv==0.19.0
python-jose==3.3.0
gunicorn==20.1.0